requirements.txt
absl-py==2.4.0
annotated-doc==0.0.4
anyio==4.12.1
astunparse==1.6.3
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cuda-bindings==12.9.4
cuda-pathfinder==1.2.2
cycler==0.12.1
deepface==0.0.98
filelock==3.20.3
fire==0.7.1
Flask==3.1.2
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.2.0
gast==0.7.0
gdown==5.2.1
google-pasta==0.2.0
grpcio==1.76.0
gunicorn==25.0.1
waitress==3.0.2
h11==0.16.0
h5py==3.15.1
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
keras==3.13.2
kiwisolver==1.4.9
libclang==18.1.1
lightdsa==0.0.3
lightecc==0.0.4
lightphe==0.0.20
lz4==4.4.5
Markdown==3.10.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mediapipe==0.10.32
ml_dtypes==0.5.4
mpmath==1.3.0
mtcnn==1.0.0
namex==0.1.0
networkx==3.6.1
numpy==2.4.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
onnx==1.20.1
onnxruntime-gpu==1.24.3
onnxslim==0.1.87
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
opt_einsum==3.4.0
optimum==2.1.0
optimum-onnx==0.1.0
optree==0.18.0
packaging==26.0
pandas==3.0.0
pillow==12.1.0
polars==1.38.1
polars-runtime-32==1.38.1
protobuf==7.34.0
psutil==7.2.2
psycopg2-binary==2.9.10
pycparser==3.0
Pygments==2.19.2
pyparsing==3.3.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
retina-face==0.0.17
rich==14.3.2
safetensors==0.7.0
scipy==1.17.1
shellingham==1.5.4
six==1.17.0
sounddevice==0.5.5
soupsieve==2.8.3
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
tf_keras==2.20.1
tokenizers==0.22.2
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
transformers==4.57.6
triton==3.6.0
typer==0.24.1
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.4.21
ultralytics-thop==2.0.18
urllib3==2.6.3
Werkzeug==3.1.5
wrapt==2.1.1
